Add qt5 build.
This is a complete redo of the qt build, using none of the old qt4 patches.
The result is a clean build with no qt source patching necessary.

macx-clang-linux/qmake.conf is added, which is generic enough to be
upstream-able. Options are added via sed, which is gross, but the alternative
would be teaching qt's configure to be aware of building for an OSX host on
Linux.

The gitian descriptor changes are untested.

TODO: Make deterministic.