dmg: add custom libdmg-hfsplus
This is a fork of upstream libdmg-hfsplus:
https://github.com/planetbeing/libdmg-hfsplus

It contains the "dmg" program which is capable of compressing an existing
dmg. Image size for current Bitcoin-Qt.dmg is reduced from ~55M to ~18M, so
it's quite necessary.

A few hacks have been added on top which fix build issues, determinism, and
a crc error. These changes are available here:
https://github.com/theuni/libdmg-hfsplus

Gitian's usage has been added, but is untested.